Split TDS client integration tests
